Make installation process aware of multiple DNAs

+++ src/shims/build.ts
@@ -3,16 +3,35 @@ import * as path from 'path'
 import {execSync} from 'child_process'
 import {bundle} from '../common'
 
-const happs = ['./src/shims/happ-data/simple-app']
+const happs = [
+  {
+    dnas: ['./src/shims/happ-data/simple-app/'],
+    ui: './src/shims/happ-data/simple-app/ui/'
+  },
+  // {
+  //   dnas: ['./src/shims/happ-data/servicelogger/']
+  // },
+  {
+    dnas: ['./src/shims/happ-data/Holo-Hosting-App/dna-src/']
+  }
+]
 
-happs.forEach(dir => {
-  try {
-    fs.unlinkSync(path.join(dir, 'ui.tar'))
-  } catch {}
+happs.forEach(happ => {
+  if (happ.ui) {
+    const dir = happ.ui
+    execSync(`cd ${dir} && hc package --strip-meta`)
+    
+    try {
+      fs.unlinkSync(path.join(dir, 'ui.tar'))
+    } catch {}
 
-  console.log('Packaging DNA...')
-  execSync(`cd ${dir} && hc package --strip-meta`)
-  console.log('Bundling UI...')
-  bundle(path.join(dir, 'ui'), path.join(dir, 'ui.tar'))
+    console.log(`Bundling UI for ${dir} ...`)
+    bundle(dir, path.join(dir, '..', 'ui.tar'))
+  }
+
+  happ.dnas.forEach(dir => {
+    console.log(`Packaging DNA for '${dir}'...`)
+    execSync(`cd ${dir} && hc package --strip-meta`)
+  })
 })
 console.log('All done!')
